_backend/requirements.txt
homeharvest
pandas
supabase
python-dotenv
requests
# ML dependencies for rent estimation
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
